pyproject.toml
[tool.poetry]
name = "dartblog"
version = "0.1.0"
description = ""
authors = ["SamsonDm <samsondm97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
pillow = "^9.4.0"
django-ckeditor = "^6.5.1"
django-simple-captcha = "^0.5.17"
django-types = "^0.17.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
